+++ lib/ferry/logic/station/station_event.dart
@@ -6,11 +6,11 @@ class StationEvent with _$StationEvent {
     @Default(false) bool isStartChange,
     required Station startStation,
     required Station endStation,
-    required Days day,
+    required Day day,
   }) = _GetEndStationsEvent;
   const factory StationEvent.changeStartStation(Station station) =
       _ChangeStartStationEvent;
   const factory StationEvent.changeEndStation(Station station) =
       _ChangeEndStationEvent;
-  const factory StationEvent.changeDay(Days day) = _ChangeDayEvent;
+  const factory StationEvent.changeDay(Day day) = _ChangeDayEvent;
 }