+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
+## [Unreleased]
+
+### For Users
+
+#### Changed
+
+ - Home page now shows icon instead of title
+
 ## [0.1.3] - 25.05.2023
 
 ### For Users